Split large visitor class
